requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiolimiter==1.2.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==24.3.0
Automat==24.8.1
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.8
constantly==15.1.0
cryptography==44.0.0
cssselect==1.2.0
dataclasses-json==0.6.7
datasets==3.2.0
defusedxml==0.7.1
dill==0.3.8
emoji==2.14.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
filelock==3.16.1
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.9.0
greenlet==3.1.1
grpcio==1.70.0
grpcio-tools==1.70.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.1
hyperframe==6.0.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
itemadapter==0.10.0
itemloaders==1.3.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.5
langchain-voyageai==0.1.4
langdetect==1.0.9
langsmith==0.3.1
lxml==4.9.1
marshmallow==3.26.0
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
nltk==3.9.1
numpy==1.26.4
olefile==0.47
ollama==0.4.7
orjson==3.10.15
packaging==24.2
pandas==2.2.3
parsel==1.9.1
pillow==11.1.0
pip==24.2
portalocker==2.10.1
propcache==0.2.1
Protego==0.3.1
protobuf==5.29.3
psutil==7.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydantic-settings==2.7.1
PyDispatcher==2.0.5
pyOpenSSL==24.3.0
pypdf==5.3.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-iso639==2025.2.8
python-magic==0.4.27
python-oxmsg==0.0.2
pytz==2024.2
PyYAML==6.0.2
qdrant-client==1.13.2
queuelib==1.7.0
RapidFuzz==3.12.1
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-toolbelt==1.0.0
Scrapy==2.12.0
service-identity==24.2.0
setuptools==75.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.37
tenacity==9.0.0
tldextract==5.1.3
tokenizers==0.21.0
tomli==2.2.1
tqdm==4.67.1
Twisted==24.11.0
typing_extensions==4.12.2
typing-inspect==0.9.0
tzdata==2025.1
unstructured==0.16.20
unstructured-client==0.30.0
urllib3==2.2.3
voyageai==0.3.2
w3lib==2.2.1
webencodings==0.5.1
wheel==0.44.0
wrapt==1.17.2
xxhash==3.5.0
yarl==1.18.3
zope.interface==7.2
zstandard==0.23.0
